(CWC-599) Feat: Keysplitting for Interactive Shell (#28)
* Abstract ShellWebsocketService into abstract base class with implementations for ssh/ssm

+ SsmShellWebsocketService will perform keysplitting protocol whereas
ssh will not.

* Refactor hpointer logic so it is managed outside of keysplitting service

+ Removes keysplitting service set expected/current hpointer methods and
instead have this managed by the websocket service

* Initial implementation of shell input/size keysplitting messages

Currently using a synchronous approach where each new input message is
only sent after previous data ack is received

* Added id token refresh logic for keysplitting. Now when the token gets refreshed ssm-shell-websocket service detects that and performs a new keysplitting handshake

* Refactor ShellState observable in shell-websocket service

Instead of emitting a state object with booleans for ready, start,
disconnect, delete instead emit different events with type. This way we
dont need to include booleans for other event types that are not
relevant.

* Add shell unattached event to websocket service + shellReattach method

Emit a shell unattached event when receiving an unrecognized synack and
stop sending any input in the websocket when not the active client. The
outer client scope should handle this by displaying a user facing
message indicating the shell input been suspended and how to resume the
shell. When resuming the shell the shellReattach method will attempt to
reperform keysplitting handshake to become the actively attached client.

* Set shell keysplitting compatible agent version to 3.0.732.16

Co-authored-by: tflpd <aflpd@bu.edu>