Create BasePolicySummary and PolicyType string union
The base policy reduces repetition. The string union allows for better
type safety. The combination of the two allows for generalized code that
operates on the shared properties (as defined in the base interface) of
a union type. Then, the union type can be discriminated by switching on
the string union allowing for access of the unshared properties without
having to do a type cast.

Requires TypeScript 4.1